Add Control Bar and Tab React components.
Using Material UI for pre-built components and icons for time
efficiency. The control bar has a  list of tabs of the different
features: Study chat bot, YouTube video recommender, PDF viewer and
annotator, Pomodoro timer and productivity tracker, and flashcards.

The navigation bar retracts to a smaller width when the mouse hovers out
of the component area.  On hover, the navigation bar expands, and the
tabs expand on hover as well. The currenly active tab changes color/size
to indicate activeness.

Currently us map function over an array of tabs, each with the
corresponding Icon and title text.

Using Tailwind to styles and customize components.