Add ReactPdf.jsx
Using react-pdf, a renderer for creating and viewing PDF files on the
browser for React apps.

Current implementation include the ability to extract the exact text
that is highlighted by the user. The text is then cleaned to that the
output text can be used for analysis or direct use by the user.'
This text extraction is accomplished by traversing through the PDF's
element siblings and children. User highlight starts and ends are
detected by calculating the difference in distance between the points
where the mouse clicked down and up. Thus, if the difference is nonzero,
a highlight is being made, accoding to the implemented logic.